Functions

Methods for Pseudodata

Base.convert — Method

convert(::Type{DataFrame}, pdata::Vector{Pseudodata})

Convert the vector of Pseudodata type to a DataFrame

ErrorsInVariablesExtremes.ensemblemean — Method

ensemblemean(pdata::Vector{Pseudodata})

Compute the ensemble mean for each year.

Parameter estimation

Extremes.fitbayes — Function

fitbayes(model::EVA; niter::Int=5000, warmup::Int=2000)::BayesianEVA

Fit the extreme value model under the Bayesian paradigm.

Methods for fitted models

Base.convert — Method

convert(::Type{MaximumLikelihoodEVA}, fm::PseudoMaximaEVA, iter::Int)

Convert a single MCMC iteration of the fm model to a pseudo MaximumLikelihoodEVA for hacking Extremes.jl methods.

Mamba.dic — Function

dic(fm::PseudoMaximaEVA)

Compute the Deviance Information Criterion (DIC) described by Gelman et al. (2013) for the PseudoMaximaEVA model fm.

Details

Reference: Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A. & Rubin, D.B. (2013). Bayesian Data Analysis (3rd ed.). Chapman and Hall/CRC. https://doi.org/10.1201/b16018

Diagnostic plots

Extremes.diagnosticplots — Function

diagnosticplots(fm::fittedEVA)

Diagnostic plots

Extremes.histplot — Function

histplot(fm::fittedEVA)

Histogram plot

Extremes.probplot — Function

probplot(fm::fittedEVA)

Probability plot

Extremes.qqplot — Function

qqplot(fm::fittedEVA)

Quantile-Quantile plot

Extremes.returnlevelplot — Function

returnlevelplot(fm::fittedEVA)

Return level plot

Types

ErrorsInVariablesExtremes.PseudoMaximaEVA — Type

PseudoMaximaEVA(pseudodata::Pseudoensemble, model::EVA, sim::Mamba.Chains)

Construct a PseudoMaximaEVA type.

Details

Encapsulates the probability distributions for each of the unobserved data for an ensemble of pseudodata, the extreme value model and a sample from the parameters posterior distribution.

TODO: Verify if all member has the same year vector.

ErrorsInVariablesExtremes.PseudoMaximaModel — Method

PseudoMaximaModel(data::Vector{Pseudodata};
    locationcov::Vector{Variable} = Vector{Variable}(),
    logscalecov::Vector{Variable} = Vector{Variable}(),
    shapecov::Vector{Variable} = Vector{Variable}())

Creates a PseudoMaximaModel structure.

ErrorsInVariablesExtremes.Pseudodata — Type

Pseudodata(name::String, value::Vector{<:UnivariateDistribution})

Construct a Pseudodata type.

Details

Encapsulates the probability distributions for each of the unobserved data. The data are not directly observed but their distributions are known. The distributions can be different for each of the data.