Functions

Methods

Distributions.logpdf — Method

logpdf(pdata::Pseudodata, y::Vector{<:Real})

Compute the log density of each of the potential data y according to the distributions in pdata.

Distributions.pdf — Method

pdf(pdata::Pseudodata, y::Vector{<:Real})

Compute the density of each of the potential data y according to the distributions in pdata.

Base.convert — Method

convert(::Type{DataFrame}, pensemble::Pseudoensemble)

Convert the Pseudoensemble type to a DataFrame

Distributions.logpdf — Method

logpdf(pensemble::Pseudoensemble, y::Vector{<:Real})

Compute the logpdf of each of the potential data y according to the distributions in pensemble.

Details

Independance is assumed between the members of the pseudoensemble.value, i.e. the sum of the logpdf of each member is taken.

Distributions.pdf — Method

pdf(pensemble::Pseudoensemble, y::Vector{<:Real})

Compute the pdf of each of the potential data y according to the distributions in pensemble.

Details

Independance is assumed between the members of the pseudoensemble.value, i.e. the product of the pdf of each member is taken.

ErrorsInVariablesExtremes.ensemblemean — Method

ensemblemean(pdata::Pseudoensemble)

Compute the ensemble mean for each year.

Types

ErrorsInVariablesExtremes.PseudoMaximaEVA — Type

PseudoMaximaEVA(pseudodata::Pseudoensemble, model::EVA, sim::Mamba.Chains)

Construct a PseudoMaximaEVA type.

Details

Encapsulates the probability distributions for each of the unobserved data for an ensemble of pseudodata, the extreme value model and a sample from the parameters posterior distribution.

TODO: Verify if all member has the same year vector.

ErrorsInVariablesExtremes.Pseudodata — Type

Pseudodata(name::String, value::Vector{<:UnivariateDistribution})

Construct a Pseudodata type.

Details

Encapsulates the probability distributions for each of the unobserved data. The data are not directly observed but their distributions are known. The distributions can be different for each of the data.

ErrorsInVariablesExtremes.Pseudoensemble — Type

Pseudoensemble(name::String, value::Vector{Pseudodata})

Construct a Pseudoensemble type.

Details

Encapsulates the probability distributions for each of the unobserved data for an ensemble of pseudodata. For the moment, independance of each Pseudodata elements is assumed.

TODO: Verify if all member has the same year vector.